feat: FileTransferService - Handle special characters in file name (#121)
FileTransferService - Handle special characters in file name

It was noticed that files containing special characters in name couldn't
be transferred via the API.
This is due to the usage of metadata that do not support special
characters.
Modify the file_transfer.proto by adding a new field file_name in Chunk
message.
The medatada with key: "file-name" becomes deprecated.